Feature/refactor string format (#16)
* Updates README

Adds --max-line-length=127 flag to flake8 example.
The default value is 80, but 127 it's more reasonable for these
days. Also, 127 is the number of columns in a GitHub editor.

* Updates to more permissive requirements

Since this is a SDK, probably it will be installed in an env
containing other packages (eg. Jupyter, scikit-learn, and others).
Having strict requirements will cause unnecessary reinstalls of
packages due to minor changes in versions (eg. minio 5.0.6->5.0.7)
**This is a test to keep these flexible versions (>=), but some
issues may occur. We'll wait and see what happens.**

* Replaces all str.format occurrencies by f-string

f-string is a new way of formatting strings since Python 3.6
It's less verbose, so code was updated to use this new syntax.